added booking adverts and some DTO

+++ GetItFreeService/src/main/java/com/getItFree/repository/AdvertRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdvertRepository extends JpaRepository<Advert, BigInteger> {
+public interface AdvertRepository extends JpaRepository<Advert, Long> {
 }
